+++ include/iap/odometry/bspline_control_window.hpp
@@ -15,6 +15,7 @@ namespace iap {
 inline constexpr std::size_t kBSplineControlPointCount = 4;
 
 gtsam::Key bspline_control_point_key(std::size_t index);
+gtsam::Key bspline_velocity_key(std::size_t index);
 
 struct BSplineControlPointState {
   std::size_t index = 0;